+++ packages/openapi2aspida/samples/openapi/api/stream/v1/stories/_storyId.ts
@@ -5,8 +5,8 @@ export interface Methods {
   get: {
     query: {
       token: string
-      offset: number
-      limit: number
+      offset?: number
+      limit?: number
     }
 
     resData: {