+++ packages/openapi2aspida/samples/openapi/api/v3/channels/_channelId/index.ts
@@ -3,11 +3,11 @@ import * as Types from '../../../../@types'
 
 export interface Methods {
   get: {
-    reqHeaders: Types.AppIdHeader & Types.AppPlatformHeader & Types.AppVersionHeader & Types.AppOrganisationToken & Types.UserInstanceIdHeader
+    reqHeaders?: Types.AppIdHeader & Types.AppPlatformHeader & Types.AppVersionHeader & Types.AppOrganisationToken & Types.UserInstanceIdHeader
 
-    query: {
-      excludeItems: boolean
-      excludePinned: boolean
+    query?: {
+      excludeItems?: boolean
+      excludePinned?: boolean
     }
   }
 }