+++ packages/openapi2aspida/samples/openapi/api/v3/channels/_channelId/news-feed.ts
@@ -3,26 +3,25 @@ import * as Types from '../../../../@types'
 
 export interface Methods {
   get: {
-    reqHeaders: Types.AppIdHeader & Types.AppPlatformHeader & Types.AppVersionHeader & Types.AppOrganisationToken & Types.UserInstanceIdHeader
+    reqHeaders?: Types.AppIdHeader & Types.AppPlatformHeader & Types.AppVersionHeader & Types.AppOrganisationToken & Types.UserInstanceIdHeader
 
-    query: {
-      timestamp: number
-      offset: number
-      limit: number
-      reverse: boolean
+    query?: {
+      timestamp?: number
+      offset?: number
+      limit?: number
+      reverse?: boolean
     }
 
     resData: {
       count: number
       data: Types.ModelCard & {
-        author: {
-          type: string
-          name: string
-          image: Types.ImageModel
+        author?: {
+          type?: string
+          name?: string
         }
-        storyId: number
-        reactions: Types.ReactionCountModel & {
-          myReaction: Types.ReactionEnumModel
+        storyId?: number
+        reactions?: Types.ReactionCountModel & {
+
         }
       }[]
     }