+++ packages/openapi2aspida/samples/openapi/api/v3/extension/audio.ts
@@ -3,7 +3,7 @@ import * as Types from '../../../@types'
 
 export interface Methods {
   post: {
-    reqHeaders: Types.AppIdHeader & Types.AppPlatformHeader & Types.AppVersionHeader & Types.AppOrganisationToken
+    reqHeaders?: Types.AppIdHeader & Types.AppPlatformHeader & Types.AppVersionHeader & Types.AppOrganisationToken
 
     resData: {
       id: string
@@ -11,7 +11,7 @@ export interface Methods {
 
     reqType: FormData
 
-    reqData: {
+    reqData?: {
       file: ArrayBuffer
     }
   }