+++ packages/openapi2aspida/samples/openapi/api/v3/organisation/users.ts
@@ -3,16 +3,16 @@ import * as Types from '../../../@types'
 
 export interface Methods {
   get: {
-    reqHeaders: Types.AppIdHeader & Types.AppPlatformHeader & Types.AppVersionHeader & Types.AppOrganisationTokenRequired
+    reqHeaders?: Types.AppIdHeader & Types.AppPlatformHeader & Types.AppVersionHeader & Types.AppOrganisationTokenRequired
 
     query: {
       search: string
-      offset: number
+      offset?: number
     }
 
     resData: {
       count: number
-      offset: number
+      offset?: number
       data: {
         email: string
         id: number